[level 2] Title: 피보나치 수, Time: 0.90 ms, Memory: 4.12 MB -BaekjoonHub

+++ 프로그래머스/2/12945. 피보나치 수/피보나치 수.cpp
@@ -5,7 +5,7 @@ using namespace std;
 
 int solution(int n) {
     int answer = 0;
-    vector<long long> fib={0,1};
+    vector<int> fib={0,1};
     for(int i = 2; i<=n; i++){
         fib.push_back((fib[i-2]%1234567)+(fib[i-1]%1234567));
     }
